requirements.txt
amqp==5.1.1
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
base58==2.1.1
billiard==4.1.0
celery==5.3.4
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coincurve==18.0.0
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.3
kombu==5.3.2
prompt-toolkit==3.0.39
pycparser==2.21
pysha3==1.0.2
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
rabbitmq==0.2.0
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
